chore: move routes under src/app and link /credential from homepage
